Move error reporting to its own module
