+++ src/lib.rs
@@ -16,6 +16,7 @@ extern crate rustc_data_structures;
 extern crate rustc_mir;
 extern crate rustc_target;
 
+mod diagnostics;
 mod eval;
 mod helpers;
 mod intptrcast;
@@ -41,6 +42,7 @@ pub use crate::shims::time::EvalContextExt as TimeEvalContextExt;
 pub use crate::shims::tls::{EvalContextExt as TlsEvalContextExt, TlsData};
 pub use crate::shims::EvalContextExt as ShimsEvalContextExt;
 
+pub use crate::diagnostics::report_err;
 pub use crate::eval::{create_ecx, eval_main, MiriConfig, TerminationInfo};
 pub use crate::helpers::EvalContextExt as HelpersEvalContextExt;
 pub use crate::machine::{
